[feat] 데미지 세팅 완료

+++ Assets/JSW/Scripts/Weapon/Projectile/Arrow.cs
@@ -49,7 +49,7 @@ public class Arrow : ProjectileBase
         }
     }
 
-    public void SetInit(Vector2 dir, int damageNum, float speedNum, float knockbackPowerNum, float scaleNum, Archer archer, bool isSkill)
+    public void SetInit(Vector2 dir, float damageNum, float speedNum, float knockbackPowerNum, float scaleNum, Archer archer, bool isSkill)
     {
         direction = dir.normalized;
         float angle = Mathf.Atan2(dir.y, dir.x) * Mathf.Rad2Deg;
