+++ Assets/JSW/Scripts/Weapon/Projectile/BlackHoleAttack.cs
@@ -8,7 +8,7 @@ public class BlackHoleAttack : ProjectileBase
     public float pullDelay = 0.3f;
     public float pullForce = 10f;
 
-    public void SetInit(Vector2 dir, int damageNum, float speedNum, float scaleNum)
+    public void SetInit(Vector2 dir, float damageNum, float speedNum, float scaleNum)
     {
         direction = dir.normalized;
         float angle = Mathf.Atan2(dir.y, dir.x) * Mathf.Rad2Deg;
@@ -60,7 +60,7 @@ public class BlackHoleAttack : ProjectileBase
                 EnemyHP hp = enemy.GetComponent<EnemyHP>();
                 if (hp != null)
                 {
-                    hp.TakeDamage(damage, ECharacterType.Magician);
+                    hp.TakeDamage((int)damage, ECharacterType.Magician);
                 }
             }
         }
