+++ Assets/JSW/Scripts/Weapon/Projectile/PriestAttack.cs
@@ -24,7 +24,7 @@ public class PriestAttack : ProjectileBase
         transform.position += (Vector3)(direction * speed * Time.deltaTime);
     }
 
-    public void SetInit(Vector2 dir, int damageNum, float speedNum, float scaleNum, Transform target = null)
+    public void SetInit(Vector2 dir, float damageNum, float speedNum, float scaleNum, Transform target = null)
     {
         direction = dir.normalized;
         float angle = Mathf.Atan2(dir.y, dir.x) * Mathf.Rad2Deg;
@@ -45,7 +45,7 @@ public class PriestAttack : ProjectileBase
         {
             var enemy = other.GetComponent<EnemyHP>();
             if (enemy != null)
-                enemy.TakeDamage(damage, ECharacterType.Priest);
+                enemy.TakeDamage((int)damage, ECharacterType.Priest);
 
             DestroyProjectile(gameObject);
         }
